Small refactoring of dl_verify parameter
