+++ oggm/cfg.py
@@ -330,6 +330,7 @@ def initialize(file=None, logging_level='INFO'):
     PARAMS['use_tar_shapefiles'] = cp.as_bool('use_tar_shapefiles')
     PARAMS['clip_mu_star'] = cp.as_bool('clip_mu_star')
     PARAMS['clip_tidewater_border'] = cp.as_bool('clip_tidewater_border')
+    PARAMS['dl_verify'] = cp.as_bool('dl_verify')
 
     # Climate
     PARAMS['baseline_climate'] = cp['baseline_climate'].strip().upper()
@@ -369,7 +370,7 @@ def initialize(file=None, logging_level='INFO'):
            'tstar_search_window', 'use_bias_for_run', 'hydro_month_sh',
            'use_intersects', 'filter_min_slope', 'clip_tidewater_border',
            'auto_skip_task', 'correct_for_neg_flux', 'filter_for_neg_flux',
-           'rgi_version',
+           'rgi_version', 'dl_verify',
            'use_shape_factor_for_inversion', 'use_rgi_area',
            'use_shape_factor_for_fluxbasedmodel', 'baseline_climate']
     for k in ltr:
@@ -414,7 +415,6 @@ def oggm_static_paths():
         config = ConfigObj()
         config['dl_cache_dir'] = os.path.join(dldir, 'download_cache')
         config['dl_cache_readonly'] = False
-        config['dl_verify'] = True
         config['tmp_dir'] = os.path.join(dldir, 'tmp')
         config['cru_dir'] = os.path.join(dldir, 'cru')
         config['rgi_dir'] = os.path.join(dldir, 'rgi')
@@ -432,7 +432,7 @@ def oggm_static_paths():
         sys.exit()
 
     # Check that all keys are here
-    for k in ['dl_cache_dir', 'dl_cache_readonly', 'dl_verify', 'tmp_dir',
+    for k in ['dl_cache_dir', 'dl_cache_readonly', 'tmp_dir',
               'cru_dir', 'rgi_dir', 'test_dir', 'has_internet']:
         if k not in config:
             raise RuntimeError('The oggm config file ({}) should have an '
@@ -466,7 +466,6 @@ def oggm_static_paths():
     # Other
     PARAMS['has_internet'] = config.as_bool('has_internet')
     PARAMS['dl_cache_readonly'] = config.as_bool('dl_cache_readonly')
-    PARAMS['dl_verify'] = config.as_bool('dl_verify')
 
     # Create cache dir if possible
     if not os.path.exists(PATHS['dl_cache_dir']):
