+++ oggm/params.cfg
@@ -47,6 +47,9 @@ use_tar_shapefiles = True
 # If you receive "Message truncated" errors from MPI, increase this
 mpi_recv_buf_size = 131072
 
+# Check for the integrity of the files OGGM downloads at run time
+dl_verify = True
+
 ### CENTERLINE determination
 
 # Decision on grid spatial resolution for each glacier
